phase-02-web-app/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
# psycopg2-binary==2.9.9  # Commented out for easier installation during testing
# Using SQLite for development/testing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
slowapi==0.1.9